Finished documentation interface
Wrote a formatting scheme for the docs output, that creates a nice
little DSL for outputting the documentation in a nice format. Still need
to let users output with JSON, but that's for another day.

Also made some small changes to config handling, with the only obvious
change being that you're now allowed to set and unset config values (to
turn off the VERBOSE setting which makes viewser output big help
messages).

This setup should be nicely extensible, and is made better by changing
things in the backend (by adding str-str dict entries in the data field
for entries lets you add stuff to the list views.