Finished FP rewrite of QS crud
Finished the FP rewrite of queryset CRUD operations. This comes with
much better error handling and output, as well as better resilience
against errors, and a nice foundation for future work.

Still need to run more tests on this to see if it's stable, but it looks
good so far!